I don't have no idea
My mod0 seems to be mostly 0s, with no useful information. It also
seems there's a .gnuid note section in the mod0. I'm parsing out the
dynstr/dynsym fields in the file header, and one of the numbers does
accuratley point to the beginning of a table of symbol names, but the
other 4 bytes point somewhere garbage (this is all rodata relative for
some reason).

Switch wiki says there's .dynamic stuff in mod0, or maybe an offset, but
I don't really get it.